pyproject.toml
[tool.poetry]
name = "aperture"
version = "0.1.0"
description = "A simple platform allowing you to prove ownership of your accounts"
authors = ["Amber Bertucci <amber@starchild.systems>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/Akarys42/aperture"
repository = "https://github.com/Akarys42/aperture"

[tool.poetry.dependencies]
python = "^3.10.0"
fastapi = "^0.86.0"
uvicorn = {extras = ["standard"], version = "^0.19.0"}
requests = "^2.28.1"
cryptography = "^38.0.3"
PyJWT = "^2.6.0"

[tool.poetry.dev-dependencies]
isort = "^5.10.1"
black = "^21.6b0"
flake8 = "^4.0.0"
flake8-annotations = "^2.6.2"
flake8-bugbear = "^22.1.11"
flake8-docstrings = "^1.6.0"
flake8-isort = "^4.1.1"
flake8-string-format = "^0.3.0"
flake8-todo = "^0.7"
pep8-naming = "^0.12.1"
pre-commit = "^2.13.0"
taskipy = "^1.8.1"

[tool.isort]
profile = "black"
line_length = 100

[tool.black]
line-length = 100

[tool.taskipy.tasks]
bot = { cmd = "python -m starbot", help = "Runs the main bot process"}
lint = { cmd = "pre-commit run --all-files", help = "Lints project files" }
precommit = { cmd = "pre-commit install", help = "Installs the pre-commit git hook" }
format = { cmd = "black --target-version py310 .", help = "Runs the black python formatter" }

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
